Add files via upload
Login Form:
The login form is a user interface component where users provide their credentials, typically an email or username along with a password, to access a secure system or application. 
Sign-Up Form:
The sign-up form is a user interface component used for user registration on a website or application.